Actually it wasn't a bug
I thought it would allow player 1 or 2 to have several turns in a row, but I realized: if a player uncovers the same spot again, it doesn't increase the number of zeroes on the board. It will still be player x's turn, which makes sense, since their turn didn't do anything.